[edit]: cast

+++ src/org/experis/train/CalcolaBiglietto.java
@@ -23,7 +23,7 @@ public class CalcolaBiglietto {
         if (age < 18) discount = 20;
         if (age > 65) discount = 40;
 
-        double priceTicket = (double) (km * priceKm) * ((100.0 - discount) / 100);
+        double priceTicket = (km * priceKm) * ((100.0 - discount) / 100);
 
         System.out.println("Price Ticket: " + priceTicket);
